Move EnsureDevice out of its folder
